Change type prise in Product class

+++ src/main/java/com/internet/shop/model/Product.java
@@ -1,13 +1,15 @@
 package com.internet.shop.model;
 
+import java.math.BigDecimal;
+
 public class Product {
     private Long id;
     private String name;
-    private double price;
+    private BigDecimal price;
 
     public Product(String name, double price) {
         this.name = name;
-        this.price = price;
+        this.price = BigDecimal.valueOf(price);
     }
 
     public Long getId() {
@@ -26,11 +28,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
